pyproject.toml
[tool.poetry]
name = "codetutor"
version = "1.o.0"
description = "A udemy clone for learning how to code."
authors = ["Naitik Mundra <mundranaitik@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.63.0"
piccolo = "^0.16.4"
piccolo-api = "^0.12.4"
piccolo-admin = "^0.11.4"
loguru = "^0.5.3"
python-dotenv = "^0.15.0"
uvicorn = {extras = ["standard"], version = "^0.12.0"}
uvloop = "^0.14.0"
passlib = "^1.7.4"
python-jose = {extras = ["cryptograpy"], version = "^3.2.0"}
pydantic = {extras = ["email"], version="^1.7"}
requests = "^2.25.1"
gunicorn = "^20.0.4"
PyVimeo = "^1.1.0"
black = "^21.6b0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
jedi = "^0.18.0"
pylint = "^2.6.0"
flake8 = "^3.8.4"
mypy = "^0.800"
ipython = "^7.21.0"
coveralls = "^3.0.1"
hypothesis = "^6.3.3"
pynvim = "^0.4.3"
neovim = "^0.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
